% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oyster_RF_package.R
\name{o_edulis_w}
\alias{o_edulis_w}
\title{Estimate Ostrea edulis wet or dry weights (both in g) from size data.}
\usage{
o_edulis_w(x, model, type, weight)
}
\arguments{
\item{x}{A vector or matrix including at least data on oyster length (in mm), depending on the model used, it will be necessary to also include data on oyster height (for models 2 and 3) and width (for model 3). Check description for more details.}

\item{model}{The model to be used to predict oyster weight. The current version has 3 models: 1) Based on oyster length; 2) based on oyster length and height, and ; 3) based on oyster length, height and width.}

\item{type}{The type of weight to be transformed. The current version allows one of three options: a) Total weight ('total'); b) Shell weight ('shell'), and; c) Soft tissue weight ('tissue').}

\item{weight}{Parameter to define if the weight to be estimated has to be wet weight ('wet') or dry weight ('dry'). Check description for more details.}
}
\description{
An application of the models originally developed by Pineda-Metz et al. (2023) for estimating total, shell and soft tissue wet weights of European flat oyster (Ostrea edulis) based on size data.

The models are based in the random forest algorithm developed by Breiman (2001). From the original models of Pineda-Metz et al. (2023) only the one for estimating soft tissue is as originally developed, whereas those for estimating total and shell wet weights are updated versions. The models have the following mean relative errors:
- Model 1 / 2 / 3 for Total Weight = 58.6% / 33.1% / 29.6%
- Model 1 / 2 / 3 for Shell Weight = 57.9% / 54.7% / 52.6%
- Model 1 / 2 / 3 for Soft Tissue Weight = 19.6% / 16.6% / 15.3%

When using the formula, check that the names of the columns/vectors from which length, width and height data will be extracted are noted as 'Length', 'Width' and 'Height' respectively.

Note that the definitions of oyster length, width and height do not follow the original description provided by Pineda-Metz et al. (2023), but the following:
- Length: Umbo hinge to longest edge.
- Width: Longest distance across the valve.
- Height: Maximum distance between external surfaces of the umbo.
}
\examples{
# Import oyster biometric data
data(oyster)

#Estimate oyster total wet weight from model based on length data only
o_edulis_w(oyster,model=1,type='total',weight='wet')

}
\references{
Breiman L (2001) Random forests. Mach Learn 45, 5-32

Pineda-Metz SEA, Merk V, Pogoda B (2023) A machine learning model and biometric transformations to facilitate European oyster monitoring. Aquat Conserv 33(7), 708-720. https://doi.org/10.1002/aqc.3912
}
\author{
Santiago E.A. Pineda-Metz
}
