% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oyster.R
\docType{data}
\name{oyster}
\alias{oyster}
\title{Oyster biometric data}
\format{
## `oyster`
A data frame with 1,401 rows and 11 columns:
\describe{
  \item{Oyster_Numbe}{An ID provided to the measured oyster, it differs from the original data set.}
  \item{Length}{Length of the oyster (in mm) measured from umbo hinge to longest edge.}
  \item{Width}{Width of the oyster (in mm) measured as longest distance across the valve.}
  \item{Height}{Height of the oyster (in mm) measured as maximum distance between external surfaces of the umbo.}
  \item{Type}{Single oysters or oyster cluster (including number of oysters): Reef_1 = One live oyster settled on a dead shell, Reef_2 = Two live oysters attached.}
  \item{Total_Wet_Wight}{Total wet weight of the oyster in g.}
  \item{Total_Dry_Wight}{Total dry weight of the oyster in g.}
  \item{Shell_Wet_Wight}{Shell wet weight of the oyster in g.}
  \item{Shell_Dry_Wight}{Shell dry weight of the oyster in g.}
  \item{Soft_Tissue_Wet_Wight}{Soft tissue wet weight of the oyster given in g.}
  \item{Soft_Tissue_Dry_Wight}{Soft tissue dry weight of the oyster given in g.}
}
}
\source{
<https://doi.pangaea.de/10.1594/PANGAEA.949238>
}
\usage{
oyster
}
\description{
A data set containing size and weight measurements of Ostrea edulis (European flat oyster).
The data set was extracted from the PANGAEA data repository and was collected within the frame of the RESTORE project.
}
\keyword{datasets}
