% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oyster_RF_package.R
\name{ww_dw}
\alias{ww_dw}
\title{Transform Ostrea edulis wet weights to dry weights.}
\usage{
ww_dw(x, weight)
}
\arguments{
\item{x}{A vector including the wet weight to be transformed.}

\item{weight}{The type of weight to be transformed. The current version allows one of three options: a) Total weight ('total'); b) Shell weight ('shell'), and; c) Soft tissue weight ('tissue').}
}
\description{
A simple way of transforming available weight data for the European flat oyster (Ostrea edulis). The transformation factors were published in Pineda-Metz et al. (2023), but will be updated when new data is available.
Note that the current version of the function only allows for input in vector form rather than a more complex input type (updates should improve this).
}
\examples{
library(oedulis)
#Oyster data from Pineda-Metz et al. 2022
data(oyster)
#Test with total wet weight data
tww=oyster$Total_Wet_Weight

ww_dw(x=tww,weight='total')

}
\references{
Pineda-Metz SEA, Merk V, Pogoda B (2023) A machine learning model and biometric transformations to facilitate European oyster monitoring. Aquat Conserv 33(7), 708-720. https://doi.org/10.1002/aqc.3912
}
\author{
Santiago E.A. Pineda-Metz
}
